% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAConSHP.R
\encoding{UTF-8}
\name{AAConSHP}
\alias{AAConSHP}
\title{AAConSHP}
\usage{
AAConSHP(rep)
}
\arguments{
\item{rep}{CHARACTER. Adresse du fichier .shp de la zone d'étude. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du fichier.}
}
\value{
La fonction renvoit un affichage des éléments.
}
\description{
La fonction \code{AAConSHP} détecte la présence d'aires d'alimentation de captage dans l'environnement immédiat d'un shapefile.
}
\details{
La fonction s'inspire du package \code{cadastreAnalysis}
}
\examples{
### Fonctionnement :
  AAConSHP(rep=F)
}
\author{
Paul CARTERON <\email{carteronpaul@gmail.com}>
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
