% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BD_PARCA.R
\encoding{UTF-8}
\name{BD_PARCA}
\alias{BD_PARCA}
\title{BD_PARCA
Chargement et export des donnees de la IGN (c) BD_PARCELLAIRE (r)}
\usage{
BD_PARCA(PARCA, rep_BDPARCA)
}
\arguments{
\item{PARCA}{sf du parcellaire cadastral. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du fichier.}

\item{rep_BDPARCA}{CHARACTER. Répertoire de la BD_PARCELLAIRE utilisée}
}
\value{
\item{ROAD_polygon}{Shapefile Vides cadatrés environnants la propriété: routes+tronçons fluviaux}  
\item{PARCELLES_polygon}{Shapefile du parcellaire cadastral global des communes} 
\item{BATICA_polygon}{Shapefile des batiments cadastrés environnants la propriété}
}
\description{
La fonction BD_PARCA charge les données de la IGN (c) BD_PARCELLAIRE (r) autour d'un parcellaire cadastral (sf) et génère un ensemble de .shp (EPSG 2154) nécessaires ou utiles à la réalisation d'une cartographie forestière ponctuelle.
}
\examples{
### Fonctionnement :
  BD_PARCA(PARCA)
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
