% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDTOPO_HYDRO.R
\encoding{UTF-8}
\name{BDTOPO_HYDRO}
\alias{BDTOPO_HYDRO}
\title{BDTOPO_HYDRO
Telechargement des donnees IGN(c) BD TOPO(r) HYDROGRAPHIE}
\usage{
BDTOPO_HYDRO(PARCA, source)
}
\arguments{
\item{PARCA}{Objet sf du parcellaire cadastral. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du fichier.}

\item{source}{CHARACTER. Source de la IGN© BD TOPO® HYDROGRAPHIE utilisée. Deux choix possibles: "IGN© BD TOPO® Hydrographie Départementale" ou "IGN© BD TOPO® Hydrographie Métropole". Si \code{FALSE}, la fonction génère une boite de dialogue de sélection.}
}
\value{
\item{HYDRO_polygon}{Objet sf ; SURFO: polygones hydrologique des surfaces hydrographiques permanentes, SURFOi: intermittentes, RESO: réservoirs d'eau}
\item{HYDRO_line}{Objet sf ; RU: lignes hydrologiques des tronçons fluviaux permanentes, RUi: intermittents, CANO: canalisations}
\item{HYDRO_point}{Objet sf ; PTSO : Points hydrologiques des points d'eau; NOMO: hydronymes}
}
\description{
La fonction \code{BDTOPO_HYDRO} charge et exporte les données IGN© BD TOPO® HYDROGRAPHIE autour d'un parcellaire cadastral (sf) et génère un ensemble d'objet sf nécessaires à la réalisation d'une cartographie forestière ponctuelle.
}
\examples{
### Fonctionnement :
  BDTOPO_HYDRO(PARCA=F, source=F)
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
