% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDTOPO_VEG.R
\encoding{UTF-8}
\name{BDTOPO_VEG}
\alias{BDTOPO_VEG}
\title{BDTOPO_VEG
Telechargement de vegetation depuis IGN(c) BD TOPO(r))}
\usage{
BDTOPO_VEG(PARCA, repBDTOPO)
}
\arguments{
\item{PARCA}{sf du parcellaire cadastral. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du shapefile}

\item{repBDTOPO}{Répertoire de la IGN© BD TOPO®. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du dossier}
}
\value{
\item{VEG_polygon}{Objet sf ; Polygones des surfaces forestières ; TYPE='VEG'}
\item{VEG_line}{Objet sf ;Lignes contours surfaces forestières ; TYPE='VEG'}
\item{VEG_point}{Objet sf ;Points natures de culture ; TYPE='VEG'}
}
\description{
La fonction \code{BDTOPO_VEG} charge les données de végétations depuis IGN© BD TOPO® autour d'un parcellaire cadastral (sf) et génère un ensemble d'objet sf nécessaires à la réalisation d'une cartographie forestière ponctuelle.
}
\examples{
### Fonctionnement :
 BDTOPO_VEG(PARCA=F, repBDTOPO=F))
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
