% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDTOPOonPARCA.R
\encoding{UTF-8}
\name{BDTOPOonPARCA}
\alias{BDTOPOonPARCA}
\title{BDTOPOonPARCA
Telechargement de données depuis IGN(c) BD TOPO(r))}
\usage{
BDTOPOonPARCA(rep, repBDTOPO)
}
\arguments{
\item{rep}{CHARACTER. Adresse du parcellaire cadastral _PARCA_polygon.shp. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du shapefile.}

\item{repBDTOPO}{Répertoire de la IGN© BD TOPO® Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du dossier}
}
\value{
\item{COMS_polygon}{Fichier shapefile ; Communes environnantes ; Tempon de 2000m autour du PARCA_polygon}
\item{COMS_line}{Fichier shapefile ; Contours des communes environnantes ; Tempon de 2000m autour du PARCA_polygon}
\item{COM_line}{Fichier shapefile ; Contours des communes environnantes ; Tempon de 500m autour du PARCA_polygon}
\item{COMS_point}{Fichier shapefile ; Centroid des communes environnantes ; Tempon de 2000m autour du PARCA_polygon}
\item{COM_point}{Fichier shapefile ; Centroid des communes environnantes ; Tempon de 500m autour du PARCA_polygon}
\item{INFRA_polygon}{Objet sf ; Infrastructures environnantes ; Y sont intégrés TYPE='BT': les batis, TYPE='CIM' : les cimetières, TYPE='SP' : les terrains de sports, TYPE='CSTsurf' : les constructions surfaciques}
\item{INFRA_line}{Objet sf ; sf Infrastructures environnantes ; Y sont intégrés TYPE='CSTline' : les constructions linéaires, TYPE='ORO': les lignes orographiques, TYPE='LE' : les lignes électriques, TYPE='VF' : les voies ferrées}
\item{ROAD_line}{Fichier shapefile ; Routes environnantes. Y sont intégrés TYPE='RN' : routes nationnales & autoroutes, TYPE='RD' : routes départementales, TYPE='RC' : routes communales, TYPE='RF' : routes empierrées/forestières, TYPE='PN' : les pistes en terrain naturel}
}
\description{
La fonction \code{BDTOPOonPARCA} télécharge des données issues d'une IGN(c) BD TOPO(r)) autour d'un parcellaire cadastral (sf) et génère un ensemble de .shp (EPSG 2154) et d'objet sf nécessaires à la réalisation d'une cartographie forestière ponctuelle.
}
\examples{
### Fonctionnement :
  BDTOPOonPARCA(rep=F, repBDTOPO=F))
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
