% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHbyDRIAS.R
\encoding{UTF-8}
\name{BHbyDRIAS}
\alias{BHbyDRIAS}
\title{BHbyDRIAS
Traitement des donnees DRIAS}
\usage{
BHbyDRIAS(txt)
}
\arguments{
\item{txt}{Répertoire du fichier \code{.txt} issu de DRIAS. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du fichier.}
}
\value{
Les fichiers sf suivant sont généré dans l'environnement
\item{DRIAS_df}{Dataframe des données brut DRIAS} 
\item{DRIAS_serie}{Dataframe des données pour création de diagramme ombrothermique}
\item{RIAS_bh}{Dataframe des données pour création de diagramme ETP}
\item{Periode}{Horizon retenu dans le code}

Les horizons retenus sont les suivants :
\enumerate{
\item H1 Horizon proche : indicateurs calculés sur la période 2021-2050
\item H2 Horizon moyen : indicateurs calculés sur la période 2041-2070
\item H3 Horizon lointain : indicateurs calculés sur la période 2071-2100
}
}
\description{
La fonction \code{BHbyDRIAS} traite les données récupérée sur le serveur DRIAS.
Procédure:
\enumerate{
  \item Aller sur \url{https://drias-prod.meteo.fr/okapi/accueil/okapiWebDrias/index.jsp} 
  \item Cliquer sur \emph{Simulations climatiques atmosphériques} / \emph{Métropole} / \emph{INDICATEUR DRIAS 2020}
  \item Sélectioner \emph{Indicateurs mensuels 'DRIAS-2020' par horizon}
  \item Dans \emph{Sélection du jeu de données }, sélectioner votre modèle. Je conseille \emph{ALADIN63_CNRM-CM5} / \emph{RCP8.5}
  \item Dans \emph{Référence temporelle}, cocher toutes les cases
  \item Dans \emph{Référence géographique}, chercher votre secteur
  \item Dans \emph{Indicateurs météorologiques}, cliquer sur "tous"
}
}
\examples{
### Fonctionnement :
   BHbyDRIAS(txt = F)
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
