% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CADbyBDPloc.R
\encoding{UTF-8}
\name{CADbyBDPloc}
\alias{CADbyBDPloc}
\title{CADbyBDPloc
Chargement et export des donnees cadastrales depuis une IGN (c) BD_PARCELLAIRE (r) locale}
\usage{
CADbyBDPloc(rep, bdp)
}
\arguments{
\item{rep}{Character. Répertoire du shapefile. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du shapefile.}

\item{bdp}{Character. Répertoire de l'IGN© BD_PARCELLAIRE®. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du dossier.}
}
\value{
\item{ROAD_polygon}{Shapefile Vides cadatrés environnants la propriété: routes+tronçons fluviaux}  
\item{PARCELLES_polygon}{Shapefile du parcellaire cadastral global des communes} 
\item{BATICA_polygon}{Shapefile des batiments cadastrés environnants la propriété}
}
\description{
La fonction \code{CADbyBDPloc} charge les données de la IGN© BD_PARCELLAIRE® autour d'un parcellaire cadastral (sf) et génère un ensemble de .shp (EPSG 2154) nécessaires ou utiles à la réalisation d'une cartographie forestière ponctuelle.
}
\examples{
### Fonctionnement :
  CADbyBDPloc(rep=F, bdp=F)
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
