% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CADbyBDPweb.R
\encoding{UTF-8}
\name{CADbyBDPweb}
\alias{CADbyBDPweb}
\title{CADbyBDPweb
Telechargement et export des donnees cadastrales depuis l' IGN (c) BD_PARCELLAIRE (r) webservice}
\usage{
CADbyBDPweb(rep)
}
\arguments{
\item{rep}{Character. Répertoire du shapefile. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du shapefile.}
}
\value{
\item{ROAD_polygon}{Shapefile Vides cadatrés environnants la propriété: routes+tronçons fluviaux}  
\item{PARCELLES_polygon}{Shapefile du parcellaire cadastral global des communes} 
\item{BATICA_polygon}{Shapefile des batiments cadastrés environnants la propriété}
}
\description{
La fonction \code{CADbyBDPweb} charge les données de la IGN (c) BD_PARCELLAIRE (r) autour d'un parcellaire cadastral (sf) et génère un ensemble de .shp (EPSG 2154) nécessaires ou utiles à la réalisation d'une cartographie forestière ponctuelle.
}
\examples{
### Fonctionnement :
  CADbyBDPweb(rep=F)
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
