% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CADbyETALAB.R
\encoding{UTF-8}
\name{CADbyETALAB}
\alias{CADbyETALAB}
\title{CADbyETALAB
Telechargement des donnees cadastrale a partir du PCI Etalab from cadastre.data.gouv.fr}
\usage{
CADbyETALAB(rep)
}
\arguments{
\item{rep}{Character. Répertoire du shapefile. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du shapefile.}
}
\value{
\item{BATICA_polygon}{Fichier shapefile ; Batiments cadastré environnants la propriété}
\item{LIEUDIT_point}{Fichier shapefile ; Lieudits environnants la propriété}
\item{ROAD_polygon}{Fichier shapefile ; Vides cadatrés environnants la propriété : routes+tronçons fluviaux}
\item{PARCELLES_polygon}{Fichier shapefile ; Parcellaire cadastral global des communes}
\item{SUBDFISC_polygon}{Fichier shapefile ; Subdivisions fiscales}
}
\description{
La fonction \code{CADbyETALAB} télécharge les données ETALAB autour d'un parcellaire cadastral (sf) et génère un ensemble de .shp (EPSG 2154) et d'objet sf nécessaires ou utiles à la réalisation d'une cartographie forestière ponctuelle.
}
\examples{
### Fonctionnement :
  CADbyETALAB(rep=F)
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
