% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAGEF.R
\encoding{UTF-8}
\name{CAGEF}
\alias{CAGEF}
\title{CAGEF
Creation de donnees pour la realisation d'un fond vectoriel}
\usage{
CAGEF(rep, CODECA)
}
\arguments{
\item{rep}{adresse du parcellaire cadastral PARCA_polygon. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du shapefile}

\item{CODECA}{Code de référence des données cadastrales à utiliser 1: Etalab, 2: IGN© BD Parcellaire®, 3: Edigeo}
}
\description{
La fonction \code{CAGEF} utilise les fonctions \code{OSMonPARCA}, \code{BDTOPOonPARCA}, \code{BDTOPO_HYDRO}, \code{BDTOPO_VEG} et \code{ETALAB}/\code{BD_PARCA}/\code{EDIGEO} pour générer un ensemble de .shp (EPSG 2154) et d'objet sf nécessaires à la réalisation d'une cartographie forestière ponctuelle.
}
\examples{
### Fonctionnement :
  CAGEF(rep=F, CODECA=1)
}
\seealso{
Les fonctions utilisées:
\code{\link{OSMonPARCA}}, \code{\link{BDTOPOonPARCA}}, \code{\link{BDTOPO_HYDRO}}, \code{\link{BDTOPO_VEG}}, \code{\link{ETALAB}}, \code{\link{BD_PARCA}}, \code{\link{EDIGEO}}
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
