% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COURBESonSHP.R
\encoding{UTF-8}
\name{COURBESonSHP}
\alias{COURBESonSHP}
\title{COURBESonSHP
Création de ligne topohgraphique}
\usage{
COURBESonSHP(shp=F, NAME=NULL, online=F)
}
\arguments{
\item{rep}{Character. Répertoire du shapefile. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du shapefile.}

\item{source_courbes}{Character. Choix de la source. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection.}
}
\description{
La fonction \code{COURBESonSHP} permet de génrer un shapefile de courbe topographique sur une emprise.
La fonction demande le département consernée.
}
\details{
La fonction télécharge les données depuis le serveur \url{https://geoservices.ign.fr/courbes-de-niveau}
}
\examples{
### Fonctionnement :
    COURBESonSHP(rep = F)
}
\author{
Paul CARTERON <\email{carteronpaul@gmail.com}>
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
