% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPonSHP.R
\encoding{UTF-8}
\name{DEPonSHP}
\alias{DEPonSHP}
\title{DEPonSHP
Recupération du département}
\usage{
DEPonSHP(shp=F)
}
\arguments{
\item{shp}{CHARACTER. Chemion du fichier shapefile d'emprise. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du fichier}
}
\description{
La fonction \code{DEPonSHP} permet de récupérer la localisation départementale d'un shapefile
}
\details{
La fonction retourne un sf des département intersectés
}
\examples{
### Fonctionnement :
    DEPonSHP(rep = F)
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
