% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lpPM.R
\encoding{UTF-8}
\name{download_lpPM}
\alias{download_lpPM}
\title{askme_lpPM
Recuperation des relevé des locaux et parcelles des personnes morales dans un département}
\usage{
askme_lpPM(rep=F, dep=F, prop=NULL)
}
\arguments{
\item{rep}{CHARACTER. Chemin du dossier contenant les txt. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du dossier}

\item{dep}{CHARACTER. Numéro du département Si \code{FALSE}, la fonction génère une boite une boite de saisie.}

\item{prop}{CHARACTER. Dénomintation du propriétaire. Si \code{NULL}, la fonction génère une boite de saisie.}
}
\description{
La fonction \code{askme_lpPM} permet d'obtenir un relevé complet d'une personne morale pour un département.
La fonction demande le département consernée et une dénomination du propriétaire.

La fonction \code{download_lpPM} permet de télécharger en amont les données 2021. Elle renvoit le répertoire de dossier contenant les txt.
}
\details{
La fonction \code{download_lpPM} télécharge les données depuis le serveur \url{https://data.economie.gouv.fr/explore/dataset/fichiers-des-locaux-et-des-parcelles-des-personnes-morales/information/}
}
\examples{
### Fonctionnement :
    monrépertoire <- download_lpPM(T)
    askme_lpPM(rep = monrépertoire, dep=F, prop=NULL)
    
### ou :
    askme_lpPM(rep = F, dep="53", prop="mon propriétaire")
    
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
