% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDIGEO.R
\encoding{UTF-8}
\name{EDIGEO}
\alias{EDIGEO}
\title{EDIGEO
Chargement et export des donnees EDIGEO du cadastre.gouv.fr}
\usage{
EDIGEO(PARCA, rep_EDIGEO)
}
\arguments{
\item{PARCA}{sf du parcellaire cadastrale.}

\item{rep_EDIGEO}{repertoire des données EDIGEO.}
}
\value{
\item{LIEUDIT_point}{Fichier shapefile ; Lieudits environnants la propriété}
\item{ROAD_polygon}{Fichier shapefile ; Vides cadatrés environnants la propriété : routes+tronçons fluviaux}
\item{BATICA_polygon}{Fichier shapefile ; Batiments cadastré environnants la propriété}
\item{PARCELLES_polygon}{Fichier shapefile ; Parcellaire cadastral global des communes}
\item{TSURF_polygon}{Fichier shapefile ; Etangs et mares cadastrées}
\item{TINE_line}{Fichier shapefile ; Lignes cadastrales : chemins, notes, etc.}
\item{TRONFLUV_polygon}{Fichier shapefile ; Ruisseaux cadastrés}
\item{COMMUNI_line}{Fichier shapefile ; Tronçons routiers inscrits au cadastre}
\item{SUBDFISC_polygon}{Fichier shapefile ; Subdivisions fiscales}
}
\description{
La fonction \code{EDIGEO} charge les données EDIGEO du cadastre.gouv.fr autour d'un parcellaire cadastral (sf) et génère un ensemble de .shp (EPSG 2154) nécessaires ou utiles à la réalisation d'une cartographie forestière ponctuelle.
}
\examples{
### Fonctionnement :
  EDIGEO(PARCA)
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
