% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOLonSHP.R
\encoding{UTF-8}
\name{GEOLonSHP}
\alias{GEOLonSHP}
\title{GEOLonSHP
Import de la BRGM (C) BD_Charm50 (r) sur une emprise shapefile}
\usage{
GEOLonSHP(shp)
}
\arguments{
\item{shp}{Adresse du \code{.shp} de la zone d'étude. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du fichier.}
}
\value{
\item{geol_polygon}{Shapefile de la géologie sur l'emprise. Généré dans le répertoire du fichier source.}
}
\description{
La fonction \code{GEOLonSHP} téléchargement la BRGM (C) BD_Charm50 (r) départementale détecté sur l'emprise du shapefile d'entrée, réalise une intersection des couches et calcul la surface SIG par entité géologique.
}
\examples{
### Fonctionnement :
  GEOLonSHP(shp = F)
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
