% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTMLtoXLSX.R
\encoding{UTF-8}
\name{HTMLtoXLSX}
\alias{HTMLtoXLSX}
\title{HTMLtoXLSX
Creation d'un .xlsx a partir de matrices cadastrales .html}
\usage{
HTMLtoXLSX(rephtml, repRdata)
}
\arguments{
\item{rephtml}{Répertoire du dossier contenant les matrices cadastrales .html  Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du dossier.}

\item{repRdata}{Répertoire du fichier .Rdata contenant les données administratives Si \code{FALSE}, la fonction \code{\link{INSEEtoRDATA}} est lancée.}
}
\value{
\item{_matrice.xlsx}{Tableur Excel. Matrice cadastrale}
\item{_subdi.xlsx}{Tableur Excel. Matrice cadastrale avec subdivision}
\item{XLSX}{Objet dataframe. Matrice cadastrale}
\item{rep.xlsx}{Character. Répertoire de sortie}
}
\description{
La fonction \code{HTMLtoXLSX} parcourt un dossier contenant des matrices cadastrales au format .html et génère un fichier .xlsx compilant l'ensemble des matrices.
}
\examples{
### Fonctionnement :
  HTMLtoXLSX(rephtml=F, repRdata=F)
}
\seealso{
L'archive .Rdata contenant les données administratives est produite par la fonction \code{\link{INSEEtoRDATA}}.
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
