% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HYDRObyBDTloc.R
\encoding{UTF-8}
\name{HYDRObyBDTloc}
\alias{HYDRObyBDTloc}
\title{HYDRObyBDTloc
Telechargement des donnees IGN(c) BD TOPO(r) HYDROGRAPHIE}
\usage{
HYDRObyBDTloc(rep, bdt)
}
\arguments{
\item{rep}{Character. Répertoire du shapefile. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du fichier.}

\item{bdt}{Character. Répertoire de la IGN© BD TOPO®. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du dossier.}
}
\value{
\item{HYDRO_polygon}{Objet sf ; SURFO: polygones hydrologique des surfaces hydrographiques permanentes, SURFOi: intermittentes, RESO: réservoirs d'eau}
\item{HYDRO_line}{Objet sf ; RU: lignes hydrologiques des tronçons fluviaux permanentes, RUi: intermittents, CANO: canalisations}
\item{HYDRO_point}{Objet sf ; PTSO : Points hydrologiques des points d'eau; NOMO: hydronymes}
}
\description{
La fonction \code{HYDRObyBDTloc} charge et exporte les données IGN© BD TOPO® HYDROGRAPHIE autour d'un parcellaire cadastral (sf) et génère un ensemble d'objet sf nécessaires à la réalisation d'une cartographie forestière ponctuelle.
}
\examples{
### Fonctionnement :
  HYDRObyBDTloc(rep=F)
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
