% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HYDRObyBDTweb.R
\encoding{UTF-8}
\name{HYDRObyBDTweb}
\alias{HYDRObyBDTweb}
\title{HYDRObyBDTweb
Telechargement des donnees IGN(c) BD TOPO(r) HYDROGRAPHIE}
\usage{
HYDRObyBDTweb(rep)
}
\arguments{
\item{rep}{Character. Répertoire du shapefile. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du fichier.}
}
\value{
\item{HYDRO_polygon}{Objet sf ; SURFO: polygones hydrologique des surfaces hydrographiques permanentes, SURFOi: intermittentes, RESO: réservoirs d'eau}
\item{HYDRO_line}{Objet sf ; RU: lignes hydrologiques des tronçons fluviaux permanentes, RUi: intermittents, CANO: canalisations}
\item{HYDRO_point}{Objet sf ; PTSO : Points hydrologiques des points d'eau; NOMO: hydronymes}
}
\description{
La fonction \code{HYDRObyBDTweb} télécharge et exporte les données IGN© BD TOPO® HYDROGRAPHIE autour d'un parcellaire cadastral (sf) et génère un ensemble d'objet sf nécessaires à la réalisation d'une cartographie forestière ponctuelle.
}
\examples{
### Fonctionnement :
  HYDRObyBDTweb(rep=F)
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
