% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFNtoSHP.R
\encoding{UTF-8}
\name{IFNtoSHP}
\alias{IFNtoSHP}
\title{IFNtoSHP
Recuperation des donnees .shp de l'IFN}
\usage{
IFNtoSHP(code)
}
\arguments{
\item{code}{CHARACTER. Code du fichier à télécharger depuis le serveur IFN}
}
\description{
La fonction \code{IGNtoSHP} permet de télécharger les fichiers shapefiles disponibles depuis le serveur de l'IFN.
}
\details{
Les codes sont les suivants: 
#' \itemize{
  \item 'RFN' Régions forestières nationales
  \item 'RFD' Régions forestières départementales
  \item 'SER' Sylvoécorégions
  \item 'SERAR' Sylvoécorégions d'alluvions récentes
}
}
\examples{
### Fonctionnement :
  SER_polygon <- IFNtoSHP(code = 'SER')
}
\references{
Les données sont explicitées sur le portail de l'IFN: \url{http://inventaire-forestier.ign.fr/carto/carto/afficherCarto/}
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
