% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INPNbyIGNweb.R
\encoding{UTF-8}
\name{INPNbyIGNweb}
\alias{INPNbyIGNweb}
\title{INPNbyIGNweb
Telechargement des zonages environnementaux a partir de l'IGN(C) web service}
\usage{
INPNbyIGNweb(shp)
}
\arguments{
\item{shp}{Character. Répertoire du shapefile. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du shapefile.}
}
\value{
La fontion retourne un ensemble de .shp dans le dossier source.
}
\description{
La fonction \code{INPNbyIGNweb} charge et exporte les zonages environnementaux intersectés sur une zone d'étude
Pour chaque enjeux, une couche .shp contenant les zones intersectant la zone d'étude est exportée.
}
\examples{
### Fonctionnement :
  INPNbyIGNweb(shp=F, rdata=F)
}
\references{
Vous pouvez retrouvé les données environnementales sur le site de l'Institut national pour la protection de la nature (INPN) : \url{https://inpn.mnhn.fr/accueil/index}
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
