% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INPNbyRdata.R
\encoding{UTF-8}
\name{INPNbyRdata}
\alias{INPNbyRdata}
\title{INPNbyRdata
Extraction des zonages environnementaux a partir d'une .Rdata}
\usage{
INPNonSHP(shp, rdata)
}
\arguments{
\item{shp}{Character. Répertoire du shapefile. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du shapefile.}

\item{rdata}{Répertoire du fichier .Rdata contenant les données INPN. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du dossier.}
}
\value{
La fontion retourne un ensemble de .shp dans le dossier source.
}
\description{
La fonction \code{INPNbyRdata} charge et exporte les zonages environnementaux intersectés sur une zone d'étude
Pour chaque enjeux, une couche .shp contenant les zones intersectant la zone d'étude est exportée.
}
\details{
La fonction réclame une archive .Rdata externe produite par la fonction \code{\link{INPNtoRDATA}}.
}
\examples{
### Fonctionnement :
  INPNbyRdata(shp=F, rdata=F)
}
\references{
Vous pouvez retrouvé les données environnementales sur le site de l'Institut national pour la protection de la nature (INPN) : \url{https://inpn.mnhn.fr/accueil/index}
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
