% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INPNtoRDATA.R
\encoding{UTF-8}
\name{INPNtoRDATA}
\alias{INPNtoRDATA}
\title{INPNtoRDATA
Creation d'une archive .Rdata des enjeux environnementaux}
\usage{
INPNtoRDATA(rep)
}
\arguments{
\item{rep}{Répertoire de sortie de l'archive .Rdata Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du dossier.}
}
\value{
\item{list_INPN}{Liste de fichiers cartographiques sf}
\item{list_INPN_NOM}{Liste des noms des fichiers cartographiques sf de la liste précédente}
}
\description{
La fonction \code{INPNtoRDATA} télécharge depuis le site de l'INPN l'ensemble des fichiers .shp des enjeux environnementaux et sauvegarde ces couches dans une archive .Rdata
}
\details{
La fonction produit une archive .Rdata nécessaire à la fonction \code{\link{INPNonSHP}}.
}
\examples{
### Fonctionnement :
  INPNtoRDATA(rep=F)
}
\references{
Vous pouvez retrouvé les données environnementales sur le site de l'Institut national pour la protection de la nature (INPN) : \url{https://inpn.mnhn.fr/accueil/index}
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
