% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INSEEtoRDATA.R
\encoding{UTF-8}
\name{INSEEtoRDATA}
\alias{INSEEtoRDATA}
\title{INSEEtoRDATA
Recuperation des donnees administratives et creation d'une archive R.data}
\usage{
INSEEtoRDATA(rep)
}
\arguments{
\item{rep}{Répertoire de sortie de l'archive .Rdata. Si \code{NULL}, les données sont justes ajoutées à l'environnement.}
}
\value{
\item{INSEE_POST}{Base officielle des codes postaux}
\item{INSEE_REGS}{Liste des régions. Codes officiels géographiques au 1er janvier 2018.}
\item{INSEE_DEPS}{Liste des départements. Codes officiels géographiques au 1er janvier 2018.}
\item{INSEE_COMS}{Liste des communes. Codes officiels géographiques au 1er janvier 2018.}
}
\description{
La fonction \code{INSEEtoRDATA} télécharge depuis la plateforme ouverte des données publiques françaises (data.gouv) et du site de l'INSEE plusieurs fichiers de référence administratives. La fonction les sauvegarde ensuite dans une archive .Rdata si un répertoire de sauvegarde est fournit.
}
\details{
La fonction produit une archive .Rdata nécessaire à la fonction \code{\link{HTMLtoXLSX}}.
}
\examples{
### Fonctionnement :
  INSEEtoRDATA(rep=NULL)
}
\references{
Les codes des champs de l'INSEE sont explicités sur le site de l'institut : \url{https://www.insee.fr/fr/information/3363504}
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
