% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRConSHP.R
\encoding{UTF-8}
\name{IRConSHP}
\alias{IRConSHP}
\title{IRConSHP
Recuperation de l'IGN (c) BD Ortho IRC (r) sur une emprise shapefile}
\usage{
IRConSHP(rep)
}
\arguments{
\item{rep}{Character. Répertoire du shapefile. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du shapefile.}
}
\value{
\item{IRC.tif}{Fichier raster .tiff ;  IRC enregistré dans le répertoire du fichier}
}
\description{
La fonction \code{IRConSHP} télécharge IGN© BD Ortho IRC® à partir d'une emprise shapefile
}
\examples{
### Fonctionnement :
  IRConSHP(rep=F)
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
