% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LINEfromPOLYGON.R
\encoding{UTF-8}
\name{LINEfromPOLYGON}
\alias{LINEfromPOLYGON}
\title{LINEfromPOLYGON
Linearisation d'un polygon}
\usage{
LINEfromPOLYGON(polygon)
}
\arguments{
\item{polygon}{Objet polygon de type POLYGON}
}
\description{
La fonction \code{LINEfromPOLYGON} convertit un polygon en ligne sans superpositions.
}
\details{
La fonction retourne un fichier sf de type LINESTRING
}
\examples{
### Fonctionnement :
    LINEfromPOLYGON(polygon)
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
