% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOADfromIGN.R
\encoding{UTF-8}
\name{LOADfromIGN}
\alias{LOADfromIGN}
\title{LOADfromIGN
Recuperation des donnees administratives et creation d'une archive R.data}
\usage{
LOADfromIGN(rep)
}
\arguments{
\item{rep}{CHARACTER. Répertoire du dossier de téléchargement. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du dossier.}
}
\description{
La fonction \code{LOADfromIGN} permet de télécharger rapidement les données de l'IGN (C).
\tabular{ll}{
  \tab >>  BD Parcellaire® \cr
  \tab >>  BD TOPO® \cr
  \tab >>  BD FORET® V2 \cr
  \tab >>  RGE ALTI 5m \cr
  \tab >>  BD ALTI® V2 25M \cr
}
La fonction demande le département consernée.
}
\details{
La fonction télécharge les données depuis le serveur \url{http://files.opendatarchives.fr/professionnels.ign.fr/}
}
\examples{
### Fonctionnement :
  LOADfromIGN(rep = F)
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
