% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/METEOFRANCEonSHP.R
\encoding{UTF-8}
\name{METEOFRANCEonSHP}
\alias{METEOFRANCEonSHP}
\title{METEOFRANCEonSHP
Climatologie METEOFRANCE sur une zone d'etude}
\usage{
METEOFRANCEonSHP(repshp)
}
\arguments{
\item{repshp}{CHARACTER. Adresse du fichier \code{.shp} de la zone d'étude. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du fichier.}
}
\description{
La fonction \code{METEOFRANCEonSHP} télécharge les données climatologiques de la station Météo-France la plus proche du centroide de la zone détude
Les informations récupérées correspondantes aux normales et records disponibles pour cette station.
}
\details{
La fonction retourne un dataframe des données METEOFRANCE. Ces renseignements sont fournies au mois et cumulés/moyennés à l'année.
}
\examples{
### Fonctionnement :
  METEOFRANCEonSHP(repshp = F)
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
