% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHonSHP.R
\encoding{UTF-8}
\name{MHonSHP}
\alias{MHonSHP}
\title{MHonSHP
Detection des monuments historiques}
\usage{
MHonSHP(rep)
}
\arguments{
\item{rep}{CHARACTER. Adresse du fichier \code{.shp}. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du fichier.}
}
\value{
La fonction renvoit un affichage des éléments.
La fonction retourne un dataframe des données METEOFRANCE. Ces renseignements sont fournies au mois et cumulés/moyennés à l'année.
}
\description{
La fonction \code{MHonSHP} détecte la présence d'un monument historique dans l'environnement immédiat d'un shapefile.
}
\details{
La fonction réclame le package \code{cadastreAnalysis}.
}
\examples{
### Fonctionnement :
  MHonSHP(rep=F)
}
\author{
Paul CARTERON <\email{carteronpaul@gmail.com}>
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
