% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAMEofSHP.R
\encoding{UTF-8}
\name{NAMEofSHP}
\alias{NAMEofSHP}
\title{NAMEofSHP
Detection du préfixe & Actualisation des repertoires SEQUOIA}
\usage{
NAMEofSHP(shp=F)
}
\arguments{
\item{shp}{CHARACTER. Chemion du fichier shapefile d'emprise. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du fichier}

\item{actu}{BOOLEAN. Si TRUE, la fonction propose l'actualisation du répertoire si l'ancienne arborescence SEQUOIA est détectée}
}
\description{
La fonction \code{NAMEofSHP} permet de récupérer le préfixe d'un shapefile SEQUOIA et d'actualiser les répertoires SEQUOIA
}
\details{
La fonction exporte dans l'environnement le préfixe et les répertoires SEQUOIA. A défaut, elle demande un nom.
}
\examples{
### Fonctionnement :
    NAMEofSHP(rep = F)
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
