% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PARCAbyBDPloc.R
\encoding{UTF-8}
\name{PARCAbyBDPloc}
\alias{PARCAbyBDPloc}
\title{PARCAbyBDPloc
Creation d'un parcellaire cadastral .shp a partir d'une IGN (c) BD parcellaire (r) locale}
\usage{
PARCAbyBDPloc(x, bdp)
}
\arguments{
\item{x}{TIBBLE. Table cadastrale SEQUOIA. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du fichier.}

\item{bdp}{Character. Répertoire de l'IGN© BD_PARCELLAIRE®. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du dossier.}
}
\value{
\item{PARCA_bdploc_sf}{Objet sf "PARCA_bdploc_sf" exporté dans l'environnement de travail. Parcellaire cadastrale.}
\item{PARCA_bdploc_errors}{Objet tibble "PARCA_bdploc_errors" exporté dans l'environnement de travail. Références non trouvées.}
}
\description{
La fonction \code{PARCAbyBDPloc} télécharge un parcellaire cadastral à partir d'une table cadastrale SEQUOIA
}
\examples{
### Fonctionnement :
   PARCAbyBDPloc(x=F, F)
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
