% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PARCAbyBDPweb.R
\encoding{UTF-8}
\name{PARCAbyBDPweb}
\alias{PARCAbyBDPweb}
\title{PARCAbyBDPweb
Creation d'un parcellaire cadastral .shp a partir de l'etalab}
\usage{
PARCAbyBDPweb(x)
}
\arguments{
\item{x}{TIBBLE. Table cadastrale SEQUOIA. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du fichier.}
}
\value{
\item{PARCA_bdpweb_sf}{Objet sf "PARCA_bdpweb_sf" exporté dans l'environnement de travail. Parcellaire cadastrale.}
\item{PARCA_bdpweb_errors}{Objet tibble "PARCA_bdpweb_errors" exporté dans l'environnement de travail. Références non trouvées.}
}
\description{
La fonction \code{PARCAbyBDPweb} télécharge un parcellaire cadastral à partir d'une table cadastrale SEQUOIA
}
\examples{
### Fonctionnement :
   PARCAbyBDPweb(x=F)
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
