% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PARCAbyETALAB.R
\encoding{UTF-8}
\name{PARCAbyETALAB}
\alias{PARCAbyETALAB}
\title{PARCAbyETALAB
Creation d'un parcellaire cadastral .shp a partir du PCI Etalab from cadastre.data.gouv.fr}
\usage{
PARCAbyETALAB(x)
}
\arguments{
\item{x}{TIBBLE. Table cadastrale SEQUOIA. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du fichier.}
}
\value{
\item{PARCA_etalab_sf}{Objet sf "PARCA_etalab_sf" exporté dans l'environnement de travail. Parcellaire cadastrale.}
\item{PARCA_etalab_errors}{Objet tibble "PARCA_etalab_errors" exporté dans l'environnement de travail. Références non trouvées.}
}
\description{
La fonction \code{PARCAbyETALAB} télécharge un parcellaire cadastral à partir d'une table cadastrale SEQUOIA
}
\examples{
### Fonctionnement :
   PARCAbyETALAB(x=F)
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
