% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PARCAtoDIAG.R
\encoding{UTF-8}
\name{PARCAtoDIAG}
\alias{PARCAtoDIAG}
\title{PARCAtoDIAG
Creation d'un ensemble de .shp a partir d'un parcellaire cadastral .shp}
\usage{
PARCAtoDIAG(rep)
}
\arguments{
\item{rep}{CHARACTER. Adresse du fichier \code{.shp}. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du fichier.
Les fichiers .shp suivants sont produits :}
}
\value{
\item{DIAG_point}{Fichier shapefile ; placettes}
\item{DIAG_line}{Fichier shapefile ; limites}
}
\description{
La fonction \code{PARCAtoDIAG} génère un ensemble de .shp (EPSG 2154) nécessaires à la réalisation d'un diagnostic forestier à partir d'un parcellaire cadastral .shp SEQUOIA (EPSG 2154).
}
\examples{
### Fonctionnement :
  PARCAtoDIAG(rep=F)
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
