% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PARCAtoUA.R
\encoding{UTF-8}
\name{PARCAtoUA}
\alias{PARCAtoUA}
\title{PARCAtoUA
Creation d'un ensemble de .shp a partir d'un parcellaire cadastral .shp}
\usage{
XLSXtoPARCA(rep)
}
\arguments{
\item{rep}{CHARACTER. Adresse du fichier \code{.shp}. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du fichier.
Les fichiers .shp suivants sont produits :}
}
\value{
\item{UA_polygon}{Fichier shapefile ; Unités d'analyse}
\item{PROP_polygon}{Fichier shapefile ; Propriété unifiée}
\item{PROP_line}{Fichier shapefile ; Contours de la propriété unifiée}
\item{PROP_point}{Fichier shapefile ; Centroide de la propriété unifiée}
}
\description{
La fonction \code{PARCAtoUA} génère un ensemble de .shp (EPSG 2154) nécessaires à la réalisation d'une cartographie forestière ponctuelle à partir d'un parcellaire cadastral .shp formatée (EPSG 2154).
}
\examples{
### Fonctionnement :
  PARCAtoUA(rep=F)
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
