% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROADtoROUTE.R
\encoding{UTF-8}
\name{ROADtoROUTE}
\alias{ROADtoROUTE}
\title{ROADtoROUTE
Conversion de ROAD vers ROUTE}
\usage{
ROADtoROUTE(road)
}
\arguments{
\item{road}{CHARACTER. Adresse du fichier \code{.shp} ROAD_polygon.shp ou ROAD_line.shp. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du fichier.}
}
\value{
Les fichiers .shp suivants sont produits :
\item{ROUTE_polygon}{Fichier shapefile ; fonds routes}
\item{ROUTE_line}{Fichier shapefile ; contours routes}
}
\description{
La fonction \code{ROADtoROUTE} génére les .shp (EPSG 2154) polygone et ligne à partir de ROAD_polygon (vides cadastraux) ou de ROAD_line (routes référencées dans l'OSM ou la BDTOPO)
}
\details{
Le shapefile road d'entrée doit disposer d'une table attributaire complétée. Le champ "TYPE" doit correspondre à la nomenclature suivante =
TYPE='RN' : routes nationnales & autoroutes, TYPE='RD' : routes départementales,  TYPE='RC' : routes revêtues non départementales, TYPE='RF' : les routes empierrées/forestières, TYPE='PN' : les pistes en terrain naturel
Dans ROAD_polygon.shp, les codes TYPE='CR' pour chemins ruraux et TYPE='CC' pour chemins privés cadastrés sont acceptés en plus des codes précédents.
}
\note{
Les fichiers ROAD_line et ROAD_polygon sont produits par \code{\link{CAGEF}}.
}
\examples{
### Fonctionnement :
  ROADtoROUTE(road = F)
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
