% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCAN25onSHP.R
\encoding{UTF-8}
\name{SCAN25onSHP}
\alias{SCAN25onSHP}
\title{SCAN25onSHP
Recuperation de l'IGN (c) Scan 25 (r) sur une emprise shapefile}
\usage{
SCAN25onSHP(rep)
}
\arguments{
\item{rep}{Character. Répertoire du shapefile. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du shapefile.}
}
\value{
\item{Scan25.tif}{Fichier raster .tiff ;  Scan 25 enregistré dans le répertoire du fichier}
}
\description{
La fonction \code{SCAN25onSHP} télécharge IGN© Scan 25® à partir d'une emprise shapefile
}
\examples{
### Fonctionnement :
  SCAN25onSHP(rep=F)
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
