% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEQUOIA.R
\encoding{UTF-8}
\name{SEQUOIA}
\alias{SEQUOIA}
\title{SEQUOIA
Appel rapide des fonctions SEQUOIA}
\usage{
SEQUOIA(enrg)
}
\arguments{
\item{enrg}{LOGICAL. Si \code{FALSE}, la fonction lance l'interface utilisateur.}
}
\description{
La fonction \code{SEQUOIA} génère une interface utilisateur permettant de faire appel aux fonctions du package.
}
\details{
La fonction ajoute les champs suivants : \code{SURF_SIG} (Surface cartographique selon R), \code{SURF_COEFF} (coefficient de correction) et \code{SURF_COR} (Surface cadastrale corrigée)
}
\examples{
### Fonctionnement :
    SEQUOIA(F)
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
