% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSPFinUA.R
\encoding{UTF-8}
\name{SSPFinUA}
\alias{SSPFinUA}
\title{SSPFinUA
Affectation arbitraire d'un code N_SSPARFOR dans UA_polygon}
\usage{
UAtoSSPF(rep)
}
\arguments{
\item{rep}{CHARACTER. Adresse du fichier \code{.shp} UA_polygon. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du fichier.}
}
\value{
\item{UA_polygon}{Fichier shapefile ; unité d'analyse corrigé}
}
\description{
La fonction \code{SSPFinUA} affecte arbitrairement un code N_SSPARFOR pour des entités de UA_polygon pour lesquels les champs PLT_TYPE, PLT_ESS, PLT_STR et PLT_TSE ont été complétés.
}
\examples{
### Fonctionnement :
    SSPFinUA(rep=F)
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
