% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/THORNTHWAITE.R
\encoding{UTF-8}
\name{THORNTHWAITE}
\alias{THORNTHWAITE}
\title{THORNTHWAITE
Calcul des ETP selon la formule de THORNTHWAITE}
\usage{
THORNTHWAITE(series, latitude, ru, snow.init, Tsnow, fr.sn.acc, snow_melt_coeff)
}
\arguments{
\item{series}{tableau des données météo}

\item{latitude}{latitude en WGS84}

\item{ru}{réserve utile}

\item{snow.init}{réservoir de neige au 1er janvier}

\item{Tsnow}{mois de neiges}

\item{fr.sn.acc}{?}

\item{snow_melt_coeff}{?}
}
\value{
tab Tableau récap des données climatologique
}
\description{
La fonction \code{THORNTHWAITE} calcule des ETP à partir des données climatologiques et de la localisation.
}
\examples{
### Fonctionnement :
    THORNTHWAITE <- function(series,  latitude, ru=80, 
                             snow.init=0, Tsnow=2, fr.sn.acc=0.95,  
                             snow_melt_coeff=1)
}
\author{
inconnu
}
