% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UAtoSSPF.R
\encoding{UTF-8}
\name{UAtoSSPF}
\alias{UAtoSSPF}
\title{UAtoSSPF
Creation d'un ensemble de .shp a partir du .shp des unites d'analyse}
\usage{
UAtoSSPF(rep)
}
\arguments{
\item{rep}{CHARACTER. Adresse du fichier \code{.shp} UA_polygon. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du fichier.}
}
\value{
\item{PF_polygon}{Fichier shapefile ; Parcellaire forestier}
\item{PF_line}{Fichier shapefile ; Contour du parcellaire forestier}
\item{SSPF_polygon}{Fichier shapefile ; Sous-parcellaire forestier cad parcellaire des unités de gestion}
\item{SSPF_line}{Fichier shapefile ; Etiquette du sous-parcellaire actualisé}
\item{AME_polygon}{Fichier shapefile ; OPTIONEL. Traitements forestiers cad aménagements}
}
\description{
La fonction \code{UAtoSSPF} génère un ensemble de .shp (EPSG 2154) nécessaires à la réalisation d'une cartographie forestière ponctuelle à partir du .shp des unités d'analyse complété (EPSG 2154).
La fonction peut également actualiser l'étiquetage du parcellaire de gestion.
}
\examples{
### Fonctionnement :
    UAtoSSPF(rep=F)
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
