% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UAtoTOPO.R
\encoding{UTF-8}
\name{UAtoTOPO}
\alias{UAtoTOPO}
\title{UAtoTOPO
Recuperation d'un modele numerique de terrain et creation des contours d'elevation pour PARCA_polygon}
\usage{
UAtoTOPO(rep)
}
\arguments{
\item{rep}{CHARACTER. Adresse du fichier \code{.shp} UA_polygon. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du fichier.}
}
\description{
La fonction \code{UAtoTOPO} lance la fonction [MNTonSHP] sur PARCA_polygon
}
\examples{
### Fonctionnement :
    UAtoTOPO(rep=F)
}
\seealso{
Voir la notice de
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
