% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UAtoUA.R
\encoding{UTF-8}
\name{UAtoUA}
\alias{UAtoUA}
\title{UAtoUA
Actualisation du .shp des unites d'analyse}
\usage{
UAtoUA(rep)
}
\arguments{
\item{rep}{CHARACTER. Adresse du fichier \code{.shp} UA_polygon. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du fichier.}
}
\value{
\item{UA_polygon}{Fichier shapefile ; unité d'analyse complété}
}
\description{
La fonction \code{UAtoUA} calcule la surface cadastrale des unités d'analyse.
}
\details{
La fonction ajoute les champs suivants : \code{SURF_SIG} Surface cartographique selon R, \code{SURF_COEFF} coefficient de correction et \code{SURF_COR} Surface cadastrale corrigée
}
\examples{
### Fonctionnement :
    UAtoUA(rep=F)
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
