% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UAtoUAS.R
\encoding{UTF-8}
\name{UAtoUAS}
\alias{UAtoUAS}
\title{UAtoUAS
Compilation des fichiers SEQUOIA}
\usage{
UAtoUAS(rep)
}
\arguments{
\item{rep}{CHARACTER. Chemin du dossier contenant les dossiers SIG. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du dossier.}

\item{data}{CHARACTER. Type de donnée SEQUOIA (Ex: UA). Si \code{FALSE}, la fonction génère une boite de dialogue}
}
\value{
\item{data_polygon}{Fichier shapefile ; donnée SEQUOIA compilées}
}
\description{
La fonction \code{UAtoUAS} permet de compiler des shapefiles SEQUOIA d'un répertoire donné
}
\examples{
### Fonctionnement :
    UAtoUAS(rep=F)
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
