% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRITE.R
\encoding{UTF-8}
\name{WRITE}
\alias{WRITE}
\title{WRITE
Export d'un .shp EPSG: 2154 ~ UTF-8}
\usage{
WRITE(sf, repout, nom)
}
\arguments{
\item{sf}{Couche sf à exporter}

\item{repout}{Répertoire d'export}

\item{nom}{Nom de la couche}
}
\description{
La fonction \code{WRITE} exporte un sf vers un répertoire de choix avec le nom choisit par l'utilisateur avec toutes les configurations voulues.
}
\examples{
### Fonctionnement :
   WRITE(PARCA, repout2, paste(NAME,"PARCA.shp",sep="_"))
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
