% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XLSXtoPARCA.R
\encoding{UTF-8}
\name{XLSXtoPARCA}
\alias{XLSXtoPARCA}
\title{XLSXtoPARCA
Creation d'un parcellaire cadastral .shp a partir d'une matrice cadastrale .xlsx}
\usage{
XLSXtoPARCA(rep)
}
\arguments{
\item{rep}{CHARACTER. Adresse du fichier \code{.xlsx}, tableur Excel contenant la matrice cadastrale Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du fichier.}
}
\value{
\item{PARCA_polygon}{Fichier shapefile ; Parcellaire cadastrale .shp EPSG 2154 pour post-traitement}
\item{PARCA-ARCHIVE_polygon}{Fichier shapefile ; Parcellaire cadastrale .shp EPSG 2154 pour archivage}
\item{PARCA_polygon}{Objet sf ; "PARCA_polygon" dans l'environnement de travail}
}
\description{
La fonction \code{XLSXtoPARCA} génère deux parcellaires cadastrals au format .shp (EPSG 2154) à partir d'une matrice cadastral formatée au format .xlsx .
}
\examples{
### Fonctionnement :
   XLSXtoPARCA(rep=F)
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
