% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INPNonSHP.R
\encoding{UTF-8}
\name{INPNonSHP}
\alias{INPNonSHP}
\title{INPNonSHP
Detection des enjeux environnementaux sur une zone d'etude}
\usage{
INPNonPARCA(repshp, repRdata)
}
\arguments{
\item{shp}{Adresse du \code{.shp} de la zone d'étude. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du fichier.}

\item{repRdata}{Répertoire du fichier .Rdata contenant les données INPN. Si \code{FALSE}, la fonction génère une boite de dialogue de sélection du dossier.}

\item{NAME}{Préfixe des fichers shapefile générés.}
}
\value{
La fontion retourne un ensemble de .shp dans le dossier source.
}
\description{
La fonction \code{INPNonSHP} recherche les intersections entre une zone d' étude (.shp quelconque) et les zonnages environnementaux référencées sur le site de l'INPN.
Pour chaque enjeux, une couche .shp contenant les zones intersectant la zone d'étude est exportée.
}
\details{
La fonction réclame une archive .Rdata externe produite par la fonction [INPNtoRDATA].
}
\examples{
### Fonctionnement :
  INPNonPARCA(shp=F, repRdata=F, NAME=NULL)
}
\references{
Vous pouvez retrouvé les données environnementales sur le site de l'Institut national pour la protection de la nature (INPN) : \url{https://inpn.mnhn.fr/accueil/index}
}
\author{
Matthieu CHEVEREAU <\email{matthieuchevereau@yahoo.fr}>
}
